refactor slack resourcesinto separate modules
