+++ busy_beaver/blueprints/integration/slack/__init__.py
@@ -1,0 +1,2 @@
+from .event_subscription import SlackEventSubscriptionResource  # noqa
+from .slash_command import SlackSlashCommandDispatchResource  # noqa
