+++ tests/blueprints/integration/slack/slash_command_test.py
@@ -1,21 +1,15 @@
-from collections import OrderedDict
-import json
-from urllib.parse import urlencode
-
 import pytest
 
-from busy_beaver.blueprints.integration.slack import (
+from busy_beaver.blueprints.integration.slack.slash_command import (
     command_not_found,
     display_help_text,
     link_github,
     next_event,
     relink_github,
 )
-from busy_beaver.config import SLACK_SIGNING_SECRET
 from busy_beaver.models import User
-from busy_beaver.decorators.verification import calculate_signature
 
-MODULE_TO_TEST = "busy_beaver.blueprints.integration.slack"
+MODULE_TO_TEST = "busy_beaver.blueprints.integration.slack.slash_command"
 
 
 @pytest.fixture
@@ -29,114 +23,6 @@ def add_user(session):
     return _add_user
 
 
-################################
-# Slack Event Subscription Tests
-################################
-@pytest.fixture
-def patched_slack(patcher):
-    def _wrapper(replacement):
-        return patcher(MODULE_TO_TEST, namespace="slack", replacement=replacement)
-
-    return _wrapper
-
-
-@pytest.fixture
-def create_slack_headers():
-    """Dictionary get sorted when we retrieve the body, account for this"""
-
-    def sort_dict(original_dict):
-        res = OrderedDict()
-        for k, v in sorted(original_dict.items()):
-            if isinstance(v, dict):
-                res[k] = dict(sort_dict(v))
-            else:
-                res[k] = v
-        return dict(res)
-
-    def wrapper(timestamp, data, is_json_data=True):
-        if is_json_data:
-            request_body = json.dumps(sort_dict(data)).encode("utf-8")
-        else:
-            request_body = urlencode(data).encode("utf-8")
-        sig = calculate_signature(SLACK_SIGNING_SECRET, timestamp, request_body)
-        return {"X-Slack-Request-Timestamp": timestamp, "X-Slack-Signature": sig}
-
-    return wrapper
-
-
-@pytest.mark.integration
-def test_slack_callback_url_verification(
-    client, session, patched_slack, create_slack_headers
-):
-    # Arrange
-    challenge_code = "test_code"
-    data = {"type": "url_verification", "challenge": challenge_code}
-    headers = create_slack_headers(100_000_000, data)
-
-    # Act
-    resp = client.post("/slack-event-subscription", headers=headers, json=data)
-
-    # Assert
-    assert resp.status_code == 200
-    assert resp.json == {"challenge": challenge_code}
-
-
-@pytest.mark.integration
-def test_slack_callback_bot_message_is_ignored(
-    mocker, client, session, patched_slack, create_slack_headers
-):
-    """Bot get notified of its own DM replies to users... ignore"""
-    # Arrange
-    slack = patched_slack(mocker.MagicMock())
-    data = {
-        "type": "unknown todo",
-        "event": {"type": "message", "subtype": "bot_message"},
-    }
-    headers = create_slack_headers(100_000_000, data)
-
-    # Act
-    resp = client.post("/slack-event-subscription", headers=headers, json=data)
-
-    # Assert
-    assert resp.status_code == 200
-    assert len(slack.mock_calls) == 0
-
-
-@pytest.mark.integration
-def test_slack_callback_user_dms_bot_reply(
-    mocker, client, session, patched_slack, create_slack_headers
-):
-    """When user messages bot, reply with help text"""
-    # Arrange
-    slack = patched_slack(mocker.MagicMock())
-    channel_id = 5
-    data = {
-        "type": "unknown todo",
-        "event": {
-            "type": "message",
-            "subtype": "not bot_message",
-            "channel_type": "im",
-            "text": "random",
-            "user": "random_user",
-            "channel": channel_id,
-        },
-    }
-    headers = create_slack_headers(100_000_000, data)
-
-    # Act
-    resp = client.post("/slack-event-subscription", headers=headers, json=data)
-
-    # Assert
-    assert resp.status_code == 200
-    assert len(slack.mock_calls) == 1
-    args, kwargs = slack.post_message.call_args
-    assert "/busybeaver help" in args[0]
-    assert kwargs["channel_id"] == channel_id
-
-
-###########################
-# Slack Slash Command Tests
-###########################
 @pytest.fixture
 def generate_slash_command_request():
     def _generate_data(
@@ -162,6 +48,9 @@ def generate_slash_command_request():
     return _generate_data
 
 
+###################
+# Integration Tests
+###################
 @pytest.mark.integration
 def test_slack_command_valid_command(
     client, create_slack_headers, generate_slash_command_request
